requirements.txt
aenum==3.1.12
allure-pytest==2.13.2
allure-python-commons==2.13.2
anyio==3.6.2
assertpy==1.1
async-generator==1.10
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
delayed-assert==0.3.6
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
paramiko==2.12.0
pexpect==4.8.0
platformdirs==2.6.2
pluggy==1.0.0
ptyprocess==0.7.0
pycparser==2.21
PyNaCl==1.5.0
PySocks==1.7.1
pytest==7.3.1
pytz==2022.7.1
reportportal-client==5.3.5
requests==2.28.1
rfc3986==1.5.0
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sshpass-wrapper==0.0.3
suds==1.1.2
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
